+++ packages/kit/src/views/Staking/components/utils.tsx
@@ -24,22 +24,3 @@ export const formatApy = (apy: string | number | undefined): string => {
   return new BigNumber(apy).decimalPlaces(2, BigNumber.ROUND_DOWN).toFixed(2);
 };
 
-export const calcPercentBalance = ({
-  balance,
-  percent,
-  decimals,
-}: {
-  balance: string;
-  percent: number;
-  decimals?: number;
-}) => {
-  const valueNumber = BigNumber(balance);
-  if (valueNumber.isZero()) {
-    return '';
-  }
-  if (percent === 100) {
-    return balance;
-  }
-  const value = valueNumber.multipliedBy(percent).dividedBy(100);
-  return decimals ? value.toFixed(decimals) : value.toFixed();
-};
